[core] SocketData moved to group_common.h
